fix: 全量 clippy 清零 — 3 workspace 包 34 warnings → 0
abacus-cli (30 warnings → 0):
- cards/*.rs: 去掉 unneeded return (×4)
- cards/render.rs: 去掉 u16→u16 冗余 cast (×3)
- components/bars.rs: format! 简化 (×2)
- components/panel.rs: format! 简化 (×2) + rsplitn 替换 (×2)
- components/mod.rs: visual_row 用 enumerate() 替代手写计数器
- markdown.rs: strip_prefix 替代手动前缀剥离 + manual arithmetic check
- md_stream.rs: impl Default for StreamingMd
- meeting_cache.rs: sort_by_key 替代 sort_by
- run.rs: 空行去除 + map_or 简化 + 裸引用优化 (×2)
- setup.rs: .first() 替代 .get(0) + str::replace 链 + format! 简化
- state/mod.rs: TaskRegistry impl Default + 裸引用优化 + &mut String→&str
- api/mod.rs: derive Default

abacus-core (1 warning → 0):
- mag_chain.rs: 嵌套 if 合并条件

abacus-ui-kit (1 warning → 0):
- theme.rs: iter().any() → contains()

测试：1544/1544 通过，0 warning，0 error

+++ pkg/crates/abacus-cli/src/tui/cards/render.rs
@@ -65,7 +65,7 @@ pub fn render_cards(f: &mut Frame, state: &AppState, area: Rect, _focus: crate::
     let is_streaming = state.cards.active_id().is_some();
     if is_streaming {
         let tick = state.anim_tick.get();
-        let shimmer_pos = ShimmerPhase::compute(tick, 50, 3500, 8, inner.width as u16);
+        let shimmer_pos = ShimmerPhase::compute(tick, 50, 3500, 8, inner.width);
         // 用 active 卡的主色 (如有), 否则 border 色
         let color = state
             .cards
@@ -149,7 +149,7 @@ pub fn render_cards(f: &mut Frame, state: &AppState, area: Rect, _focus: crate::
                 // 计算 shimmer 位置 (仅 active 卡)
                 let shimmer_pos = if state.cards.active_id() == Some(id) {
                     let tick = state.anim_tick.get();
-                    ShimmerPhase::compute(tick, 50, 3500, 8, inner.width as u16)
+                    ShimmerPhase::compute(tick, 50, 3500, 8, inner.width)
                 } else {
                     -999 // sentinel: 关闭 shimmer
                 };
@@ -176,7 +176,7 @@ pub fn render_cards(f: &mut Frame, state: &AppState, area: Rect, _focus: crate::
         // 计算 shimmer 位置 (仅 active 卡)
         let shimmer_pos = if state.cards.active_id() == Some(id) {
             let tick = state.anim_tick.get();
-            ShimmerPhase::compute(tick, 50, 3500, 8, inner.width as u16)
+            ShimmerPhase::compute(tick, 50, 3500, 8, inner.width)
         } else {
             -999 // sentinel: 关闭 shimmer
         };
